Game:GameWindow: Refactor code
